// src/api-client/models/PricingItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface PricingItem
 */
export interface PricingItem {
  /**
   *
   * @type {Date}
   * @memberof PricingItem
   */
  validFrom: Date;
  /**
   *
   * @type {Date}
   * @memberof PricingItem
   */
  validTo: Date;
  /**
   *
   * @type {number}
   * @memberof PricingItem
   */
  price: number;
}

/**
 * Check if a given object implements the PricingItem interface.
 */
export function instanceOfPricingItem(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "validFrom" in value;
  isInstance = isInstance && "validTo" in value;
  isInstance = isInstance && "price" in value;

  return isInstance;
}

export function PricingItemFromJSON(json: any): PricingItem {
  return PricingItemFromJSONTyped(json, false);
}

export function PricingItemFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PricingItem {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    validFrom: new Date(json["validFrom"]),
    validTo: new Date(json["validTo"]),
    price: json["price"],
  };
}

export function PricingItemToJSON(value?: PricingItem | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    validFrom: value.validFrom.toISOString(),
    validTo: value.validTo.toISOString(),
    price: value.price,
  };
}